+++ core/src/main/java/react4j/core/ComponentPhase.java
@@ -33,7 +33,6 @@ enum ComponentPhase
    * The component is potentially re-rendered due to updates.
    * The component goes through the steps:
    * <ul>
-   * <li>{@link Component#componentWillReceiveProps(JsPropertyMap)} if props were provided. i.e. the parent component was re-rendered.</li>
    * <li>{@link Component#shouldComponentUpdate(JsPropertyMap, JsPropertyMap)} and if the component returns false the update will be aborted.</li>
    * <li>{@link Component#render()}.</li>
    * <li>{@link Component#componentDidUpdate(JsPropertyMap, JsPropertyMap)}.</li>